requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pytesseract==0.3.13
tesseract==0.1.3
